Readme: Make it clearer that it needs an array
Thanks for this library! 

I had a gotcha where I was passing an object in instead of an array and couldn't work out what I was doing wrong.

I've suggested an update to the Readme to make that clearer. I ended up delving into the source code to understand what was going on and this could save others similar issues.